+++ src/App.tsx
@@ -1,7 +1,7 @@
 import './App.scss';
-import { Footer } from './modules/Footer';
-import { Header } from './modules/Header';
 import { HomePage } from './modules/HomePage';
+import { Footer } from './shared/Footer';
+import { Header } from './shared/Header';
 
 export const App = () => {
   return (
