+++ src/Root.tsx
@@ -1,7 +1,7 @@
 import { HashRouter as Router, Routes, Route } from 'react-router-dom';
 import { App } from './App';
-import { BurgerProvider } from './modules/shared/context/BurgerContext';
-import { ProductProvider } from './modules/shared/context/ProductsContext';
+import { BurgerProvider } from './shared/context/BurgerContext';
+import { ProductProvider } from './shared/context/ProductsContext';
 
 export const Root = () => (
   <ProductProvider>
