+++ src/modules/HomePage/Components/HotPrices/HotPrices.tsx
@@ -1,37 +1,14 @@
-import classNames from 'classnames';
-import s from './HotPrices.module.scss';
-import { ProductSlider } from '../ProductSlider';
-import { useState } from 'react';
+import { useContext } from 'react';
+import { ProductSlider } from '../../../../shared/ProductSlider';
+import { ProductContext } from '../../../../shared/context/ProductsContext';
 
 export const HotPrices = () => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const nextCard = () => {
-    setCurrentIndex(prev => prev + 1);
-  };
-
-  const prevCard = () => {
-    setCurrentIndex(prev => prev - 1);
-  };
+  const { products } = useContext(ProductContext);
+  const title = 'Hot prices';
 
   return (
     <>
-      <div
-        className={classNames(s.title__wrapper, 'container', 'block-margin')}
-      >
-        <div className={s.title}>
-          <h2>Hot prices</h2>
-        </div>
-        <div className={s.title__buttons}>
-          <button className={s.title__buttons_prev} onClick={prevCard}>
-            <img src="./img/icons/prev.png" alt="previous" />
-          </button>
-          <button className={s.title__buttons_next} onClick={nextCard}>
-            <img src="./img/icons/next.png" alt="next" />
-          </button>
-        </div>
-      </div>
-      <ProductSlider currentIndex={currentIndex} />
+      <ProductSlider title={title} products={products} />
     </>
   );
 };
