+++ src/modules/HomePage/Components/NewModelsSlider/NewModelsSlider.tsx
@@ -1,41 +1,14 @@
-import classNames from 'classnames';
-import s from './NewModelsSlider.module.scss';
-import { useState } from 'react';
-import { ProductSlider } from '../ProductSlider';
+import { useContext } from 'react';
+import { ProductContext } from '../../../../shared/context/ProductsContext';
+import { ProductSlider } from '../../../../shared/ProductSlider';
 
 export const NewModelsSlider = () => {
-  const [currentIndex, setCurrentIndex] = useState(0);
-
-  const nextCard = () => {
-    setCurrentIndex(prev => prev + 1);
-  };
-
-  const prevCard = () => {
-    setCurrentIndex(prev => prev - 1);
-  };
+  const { products } = useContext(ProductContext);
+  const title = 'Brand new models';
 
   return (
-    <div className="block-margin">
-      <div className={classNames(s.title__wrapper, 'container')}>
-        <div className={s.title}>
-          <h2>Brand new models</h2>
-        </div>
-        <div className={s.title__buttons}>
-          <button
-            className={classNames(s.title__buttons_prev, {
-              [s.disable]: currentIndex === 0,
-            })}
-            onClick={prevCard}
-            disabled={currentIndex === 0}
-          >
-            <img src="./img/icons/prev.png" alt="previous card" />
-          </button>
-          <button className={s.title__buttons_next} onClick={nextCard}>
-            <img src="./img/icons/next.png" alt="next card" />
-          </button>
-        </div>
-      </div>
-      <ProductSlider currentIndex={currentIndex} />
-    </div>
+    <>
+      <ProductSlider title={title} products={products} />
+    </>
   );
 };
