--- src/modules/HomePage/Components/ProductSlider/ProductSlider.module.scss
@@ -1,11 +1,0 @@
-.cards_slider {
-  overflow-x: hidden;
-}
-
-.cards_slide {
-  display: flex;
-  flex-wrap: nowrap;
-  column-gap: 16px;
-
-  transition: transform 0.3s ease-in-out;
-}
