+++ src/shared/Footer/Footer.module.scss
