+++ src/shared/Footer/index.ts
