+++ src/shared/Header/Components/RightButtons/RightButtons.tsx
@@ -1,7 +1,7 @@
 import React, { useContext } from 'react';
 import s from './RightButtons.module.scss';
 import classNames from 'classnames';
-import { BurgerContext } from '../../../shared/context/BurgerContext';
+import { BurgerContext } from '../../../context/BurgerContext';
 
 type Props = {
   burgerMenu?: boolean;
