+++ src/shared/Header/Header.module.scss
@@ -1,5 +1,6 @@
 @import '../../style/main';
 
+
 .header {
   // display: none !important; // important
   position: fixed;
