+++ src/shared/Header/Header.tsx
@@ -2,7 +2,7 @@ import { NavMenu } from './Components/NavMenu';
 import { RightButtons } from './Components/RightButtons';
 import s from './Header.module.scss';
 import { BurgerMenu } from './Components/BurgerMenu';
-import { Logo } from '../shared/Logo';
+import { Logo } from '../Logo';
 
 export const Header = () => {
   return (
