+++ src/shared/Logo/Logo.module.scss
@@ -1,4 +1,4 @@
-@import '../../../style/main';
+@import '../../style/main';
 
 .nav {
   &__logo_wrapper{
