+++ src/shared/Logo/index.ts
