+++ src/shared/ProductSlider/ProductSlider.module.scss
@@ -1,0 +1,37 @@
+@import '../../style/main';
+
+.title {
+  @extend %title-basic;
+
+  &__wrapper {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  &__buttons {
+    display: flex;
+    column-gap: 16px;
+
+    &_prev {
+      @extend %button-slider;
+
+    }
+
+    &_next {
+      @extend %button-slider;
+
+    }
+  }
+}
+
+.cards_slider {
+  overflow-x: hidden;
+}
+
+.cards_slide {
+  display: flex;
+  flex-wrap: nowrap;
+  column-gap: 16px;
+
+  transition: transform 0.3s ease-in-out;
+}
