+++ src/shared/ProductSlider/components/ProductCards/ProductCards.module.scss
@@ -1,4 +1,4 @@
-@import '../../../style/main';
+@import '../../../../style/main';
 
 .card {
   display: flex;
