+++ src/shared/ProductSlider/components/ProductCards/index.ts
