+++ src/shared/ProductSlider/index.ts
